Refactor `check config` tests to highlight how they depend on config
Most of the work is being done in that package, so be clear that the hub package
is just handing that off.

The config class needs to have pointers so that we can spy on function calls
with a counter in the tests. We're spying rather than comparing responses just
to not over-do the reconstruction of the JSON logic.

Signed-off-by: Shoaib Lari <slari@pivotal.io>